requirements.txt
langchain
langchain-community
langchain-ollama
python-dotenv
pandas
streamlit